GH-1245 Refactor some functionality into auto_bp_peering.hpp
